refactor: adjust userRating type and imports in MyRatingViewModel and related files

+++ domain/src/main/java/com/baghdad/domain/model/RatedMedia.kt
@@ -2,7 +2,7 @@ package com.baghdad.domain.model
 
 data class RatedMedia(
     val id: Long,
-    val userRating: Double?,
+    val userRating: Int?,
     val posterImageURL: String,
     val contentType: ContentType
 ) {
